hello

+++ glap/src/gldemo/learnOpengl/chapt7/2dimg.vert
@@ -1,7 +1,7 @@
 #version 330 core
 layout (location = 0) in vec3 position;
 layout (location = 1) in vec2 texCoord;
-layout (location = 2) in int textureIndex;
+layout (location = 2) in float textureIndex;
 out vec2 TexCoord;
 flat out float ourTextureIndex;
 void main()
